requirements_train.txt
Django==4.2.7
transformers==4.35.2
torch==2.1.1
Pillow==10.1.0
requests==2.31.0
pandas==2.1.3
scikit-learn==1.3.2
tqdm==4.66.1
accelerate==0.24.1
